requirements.txt
flask
requests
autotraders
Pillow
